Changed all Ids to Integer

+++ src/main/java/com/dasgupta/careercompass/bookmark/BookmarkRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
+public interface BookmarkRepository extends JpaRepository<Bookmark, Integer> {
     Optional<Bookmark> findByCandidateAndJob(Candidate candidate, Job job);
 
     List<Bookmark> findByCandidate(Candidate candidate);
