+++ src/main/java/com/dasgupta/careercompass/questionnaire/QuestionnaireRepository.java
@@ -2,5 +2,5 @@ package com.dasgupta.careercompass.questionnaire;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionnaireRepository extends JpaRepository<Questionnaire, Long> {
+public interface QuestionnaireRepository extends JpaRepository<Questionnaire, Integer> {
 }
