+++ src/main/java/com/dasgupta/careercompass/questionnaire/question/QuestionRepository.java
@@ -2,5 +2,5 @@ package com.dasgupta.careercompass.questionnaire.question;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question, Long> {
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
 }
